fixed error

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -3,7 +3,7 @@
 DELIMITER //
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-    DECLARE average_score FLOAT;
+    DECLARE average_score DECIMAL(10, 2);
 
     SELECT AVG(score) INTO average_score FROM corrections
     WHERE user_id = user_id;
